Factor out status codes
