--- src/ingest_validation_tools/argparse_types.py
@@ -1,12 +1,0 @@
-import os
-
-
-class ShowUsageException(Exception):
-    pass
-
-
-def dir_path(s):
-    if os.path.isdir(s):
-        return s
-    else:
-        raise ShowUsageException(f'"{s}" is not a directory')
