+++ src/validate_samples.py
@@ -6,13 +6,7 @@ import inspect
 
 from ingest_validation_tools.error_report import ErrorReport
 from ingest_validation_tools.samples import Samples
-from ingest_validation_tools.argparse_types import ShowUsageException
-
-
-VALID_STATUS = 0
-BUG_STATUS = 1
-ERROR_STATUS = 2
-INVALID_STATUS = 3
+from ingest_validation_tools.cli_utils import ShowUsageException, exit_codes
 
 
 def make_parser():
@@ -20,10 +14,10 @@ def make_parser():
         description='Validate a HuBMAP Sample metadata TSV.',
         epilog=f'''
 Exit status codes:
-  {VALID_STATUS}: Validation passed
-  {BUG_STATUS}: Unexpected bug
-  {ERROR_STATUS}: User error
-  {INVALID_STATUS}: Validation failed
+  {exit_codes.VALID_STATUS}: Validation passed
+  {exit_codes.BUG_STATUS}: Unexpected bug
+  {exit_codes.ERROR_STATUS}: User error
+  {exit_codes.INVALID_STATUS}: Validation failed
 ''')
     parser.add_argument(
         '--path', required=True,
@@ -53,7 +47,7 @@ def main():
     errors = samples.get_errors()
     report = ErrorReport(errors)
     print(getattr(report, args.output)())
-    return INVALID_STATUS if errors else VALID_STATUS
+    return exit_codes.INVALID_STATUS if errors else exit_codes.VALID_STATUS
 
 
 if __name__ == "__main__":
@@ -62,5 +56,5 @@ if __name__ == "__main__":
     except ShowUsageException as e:
         print(parser.format_usage(), file=sys.stderr)
         print(e, file=sys.stderr)
-        exit_status = ERROR_STATUS
+        exit_status = exit_codes.ERROR_STATUS
     sys.exit(exit_status)
