+++ src/validate_upload.py
@@ -8,8 +8,7 @@ import inspect
 
 from ingest_validation_tools.error_report import ErrorReport
 from ingest_validation_tools.upload import Upload
-from ingest_validation_tools import argparse_types
-from ingest_validation_tools.argparse_types import ShowUsageException
+from ingest_validation_tools.cli_utils import ShowUsageException, exit_codes, dir_path
 from ingest_validation_tools.check_factory import cache_path
 
 directory_schemas = sorted({
@@ -17,12 +16,6 @@ directory_schemas = sorted({
     (Path(__file__).parent / 'ingest_validation_tools' /
      'directory-schemas').glob('*.yaml')
 })
-
-
-VALID_STATUS = 0
-BUG_STATUS = 1
-ERROR_STATUS = 2
-INVALID_STATUS = 3
 
 
 def make_parser():
@@ -43,10 +36,10 @@ Typical usage:
   extra parameters.
 
 Exit status codes:
-  {VALID_STATUS}: Validation passed
-  {BUG_STATUS}: Unexpected bug
-  {ERROR_STATUS}: User error
-  {INVALID_STATUS}: Validation failed
+  {exit_codes.VALID_STATUS}: Validation passed
+  {exit_codes.BUG_STATUS}: Unexpected bug
+  {exit_codes.ERROR_STATUS}: User error
+  {exit_codes.INVALID_STATUS}: Validation failed
 ''',
         formatter_class=argparse.RawDescriptionHelpFormatter)
 
@@ -54,7 +47,7 @@ Exit status codes:
 
     mutex_group = parser.add_mutually_exclusive_group(required=True)
     mutex_group.add_argument(
-        '--local_directory', type=argparse_types.dir_path,
+        '--local_directory', type=dir_path,
         metavar='PATH',
         help='Local directory to validate')
     mutex_group.add_argument(
@@ -168,7 +161,7 @@ def main():
     errors = upload.get_errors()
     report = ErrorReport(errors)
     print(getattr(report, args.output)())
-    return INVALID_STATUS if errors else VALID_STATUS
+    return exit_codes.INVALID_STATUS if errors else exit_codes.VALID_STATUS
 
 
 if __name__ == "__main__":
@@ -177,5 +170,5 @@ if __name__ == "__main__":
     except ShowUsageException as e:
         print(parser.format_usage(), file=sys.stderr)
         print(e, file=sys.stderr)
-        exit_status = ERROR_STATUS
+        exit_status = exit_codes.ERROR_STATUS
     sys.exit(exit_status)
